feat : "image 변경 시 storage가 아닌 다른 곳일시 삭제안함"

+++ src/main/java/com/mayo/client/mayoclientapi/application/service/UserService.java
@@ -170,7 +170,7 @@ public class UserService {
                         ErrorStatus.toErrorStatus("해당하는 유저가 없습니다.", 404, LocalDateTime.now())
                 ));
 
-        if(user.getPhotoUrl() != null) {
+        if(user.getPhotoUrl() != null && storageService.isFirebaseStorageUrl(user.getPhotoUrl())) {
             storageService.deleteFirebaseBucket(user.getPhotoUrl());
         }
